// front/src/shared/api/orval/models/subscriptionMember.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * AI manager
 * хуйня
 * OpenAPI spec version: 1.0.0
 */
import type { Subscription } from './subscription';
import type { User } from './user';

export interface SubscriptionMember {
  /** Unique identifier for the subscription member */
  id: number;
  /** The subscription this member belongs to */
  subscription: Subscription;
  /** The user who is a member of the subscription */
  user: User;
  /** Indicates if the user is the owner of the subscription */
  isOwner: boolean;
  /** Date when the member was added */
  addedAt: string;
}